[25306][25284] Allow sticky table headers only in newer chrome versions
This PR:

* Adds a browser version testing package 'bowser'
* Refactors the previous browser checking (unsupported browsers) with it
* Adds a `-supports-sticky-headers` class to HTML when chrome version >= 58 is detected
