+++ app/assets/javascripts/application.js.erb
@@ -45,8 +45,6 @@
 //= require settings
 //= require modal
 //= require tab_handling
-//= require top-shelf
-//= require unsupported-browsers
 //= require autocomplete_textareas
 //= require openproject_plugins
 //= require versions
